// src/model/User.ts
interface User {
    id_usuario: number;
    nome_usuario: string;
    genero_usuario: string;
    estado_civil: string;
    profissao_usuario: string;
    renda_usuario: number;
    etnia_usuario: string;
    data_usuario: string;
    cpf_usuario: string;
    rg_usuario: string;
    cep_usuario: string;
    endereco_usuario: string;
    telefone_usuario: string;
    email_usuario: string;
    senha_usuario: string;
    descricao_usuario: string;
}

export default User;